requirements.txt
flask
requests
dnspython